Sync GitHub Repository with latest Changes
- Begin name change DashToHls -> UDT
- Create new UDT API Header file to replace previous header
- Include sample test content
- Add Tests
- Add new boxes (cbmp, emsg, prft, sbgp, sgpd and more)
- Numerous optimizations and memory leak fixes

- TODO:
- Completely remove DashToHls naming scheme
- Add more unit tests for new boxes

+++ library/dash/saio_contents.h
@@ -41,7 +41,7 @@ class SaioContents : public FullBoxContents {
   uint32_t get_aux_info_type() {return aux_info_type_;}
   uint32_t get_aux_info_type_parameter() {return aux_info_type_parameter_;}
 
-  const std::vector<uint32_t>& get_offsets() const {return offsets_;}
+  const std::vector<uint64_t>& get_offsets() const {return offsets_;}
 
   virtual std::string PrettyPrint(std::string indent) const;
   virtual std::string BoxName() const {
@@ -57,7 +57,7 @@ class SaioContents : public FullBoxContents {
  private:
   uint32_t aux_info_type_;
   uint32_t aux_info_type_parameter_;
-  std::vector<uint32_t> offsets_;
+  std::vector<uint64_t> offsets_;
 };
 }  //  namespace dash2hls
